+++ webknossos/tests/dataset/test_dataset_add_remote_mag_and_layer.py
@@ -1,5 +1,4 @@
 import itertools
-import os
 from collections.abc import Iterable, Iterator
 from pathlib import Path
 
@@ -34,9 +33,7 @@ def test_add_remote_mags_from_mag_view(
     sample_remote_dataset: Dataset,
     sample_layer_and_mag_name: Iterable[tuple[str, str]],
 ) -> None:
-    remote_dataset = Dataset.open_remote(
-        "l4_sample", "Organization_X", os.getenv("WK_TOKEN")
-    )
+    remote_dataset = Dataset.open_remote("l4_sample", "Organization_X")
     sample_remote_mags = [
         remote_dataset.get_layer(layer).get_mag(mag)
         for layer, mag in sample_layer_and_mag_name
@@ -64,9 +61,7 @@ def test_add_remote_mags_from_path(
     sample_remote_dataset: Dataset,
     sample_layer_and_mag_name: Iterable[tuple[str, str]],
 ) -> None:
-    remote_dataset = Dataset.open_remote(
-        "l4_sample", "Organization_X", os.getenv("WK_TOKEN")
-    )
+    remote_dataset = Dataset.open_remote("l4_sample", "Organization_X")
     sample_remote_mags = [
         remote_dataset.get_layer(layer).get_mag(mag)
         for layer, mag in sample_layer_and_mag_name
@@ -93,9 +88,7 @@ def test_add_remote_mags_from_path(
 
 
 def test_ref_layer_from_object(sample_remote_dataset: Dataset) -> None:
-    remote_dataset = Dataset.open_remote(
-        "l4_sample", "Organization_X", os.getenv("WK_TOKEN")
-    )
+    remote_dataset = Dataset.open_remote("l4_sample", "Organization_X")
     sample_remote_layer = list(remote_dataset.layers.values())
     for layer in sample_remote_layer:
         assert is_remote_path(layer.path), "Remote mag does not have remote path."
@@ -108,9 +101,7 @@ def test_ref_layer_from_object(sample_remote_dataset: Dataset) -> None:
 
 
 def test_ref_layer_from_path(sample_remote_dataset: Dataset) -> None:
-    remote_dataset = Dataset.open_remote(
-        "l4_sample", "Organization_X", os.getenv("WK_TOKEN")
-    )
+    remote_dataset = Dataset.open_remote("l4_sample", "Organization_X")
     sample_remote_layer = list(remote_dataset.layers.values())
     for layer in sample_remote_layer:
         assert is_remote_path(layer.path), "Remote mag does not have remote path."
@@ -126,9 +117,7 @@ def test_ref_layer_non_public(tmp_path: Path) -> None:
     dataset = Dataset.open("testdata/simple_zarr3_dataset").copy_dataset(
         tmp_path / "simple_zarr3_dataset"
     )
-    remote_dataset = Dataset.open_remote(
-        "l4_sample", "Organization_X", os.getenv("WK_TOKEN")
-    )
+    remote_dataset = Dataset.open_remote("l4_sample", "Organization_X")
     remote_dataset.is_public = False
     dataset.add_layer_as_ref(remote_dataset.get_layer("segmentation"), "segmentation")
 
@@ -143,9 +132,7 @@ def test_ref_layer_non_public(tmp_path: Path) -> None:
 
 def test_shallow_copy_remote_layers(tmp_path: Path) -> None:
     dataset = Dataset(tmp_path / "origin", voxel_size=(10, 10, 10))
-    remote_dataset = Dataset.open_remote(
-        "l4_sample", "Organization_X", os.getenv("WK_TOKEN")
-    )
+    remote_dataset = Dataset.open_remote("l4_sample", "Organization_X")
     dataset.add_layer_as_ref(remote_dataset.get_layer("color"), "color")
     dataset_copy = dataset.shallow_copy_dataset(tmp_path / "copy")
     data = dataset_copy.get_layer("color").get_mag("16-16-4").read()