+++ webknossos/webknossos/client/api_client/wk_api_client.py
@@ -44,10 +44,8 @@ class WkApiClient(AbstractApiClient):
         base_wk_url: str,
         timeout_seconds: float,
         headers: dict[str, str] | None = None,
-        webknossos_api_version: int = 9,
     ):
         super().__init__(timeout_seconds, headers)
-        self.webknossos_api_version = webknossos_api_version
         self.base_wk_url = base_wk_url
 
     @property