+++ webknossos/Changelog.md
@@ -31,10 +31,12 @@ For upgrade instructions, please check the respective _Breaking Changes_ section
     - `my_task.update()` to update a `Task` object.
     - `my_task.delete()` to delete a `Task` object.
 - Added `TaskExperience` class to store domain and value pair.
+- The libs client now uses versioned datastore routes for improved compatibility across webknossos versions. [#1328](https://github.com/scalableminds/webknossos-libs/pull/1328)
 
 ### Changed
 
 ### Fixed
+- Fixed the tests for the newest WEBKNOSSOS version. [#1328](https://github.com/scalableminds/webknossos-libs/pull/1328)
 
 
 ## [2.4.1](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.4.1) - 2025-07-03